Moved CustomizableSkeleton to the main folder
