+++ Clients/src/presentation/pages/ProjectView/ProjectFrameworks/index.tsx
@@ -4,7 +4,7 @@ import TabList from "@mui/lab/TabList";
 import TabPanel from "@mui/lab/TabPanel";
 import TabContext from "@mui/lab/TabContext";
 import { tabStyle, tabPanelStyle } from "../V1.0ProjectView/style";
-import CustomizableSkeleton from "../../../vw-v2-components/Skeletons";
+import CustomizableSkeleton from "../../../components/Skeletons";
 import ComplianceTracker from "../../../pages/ComplianceTracker/1.0ComplianceTracker";
 import { Project } from "../../../../domain/types/Project";
 import { Framework } from "../../../../domain/types/Framework";
@@ -63,9 +63,8 @@ const ProjectFrameworks = ({
   const [hasInitialized, setHasInitialized] = useState(false);
   const [searchParams, setSearchParams] = useSearchParams();
 
-  const { changeComponentVisibility } =
-    useContext(VerifyWiseContext);
-    const { userRoleName } = useAuth();
+  const { changeComponentVisibility } = useContext(VerifyWiseContext);
+  const { userRoleName } = useAuth();
 
   const { refs, allVisible } = useMultipleOnScreen<HTMLElement>({
     countToTrigger: 1,
@@ -76,7 +75,10 @@ const ProjectFrameworks = ({
 
   // Filter out organizational frameworks
   const nonOrganizationalFrameworks = useMemo(
-    () => allFrameworks.filter((framework: Framework) => !framework.is_organizational),
+    () =>
+      allFrameworks.filter(
+        (framework: Framework) => !framework.is_organizational
+      ),
     [allFrameworks]
   );
