+++ Clients/src/presentation/pages/ProjectView/V1.0ProjectView/Overview/index.tsx
@@ -8,9 +8,9 @@ import DescriptionCard from "../../../../components/Cards/DescriptionCard";
 import TeamCard from "../../../../components/Cards/TeamCard";
 import { Project } from "../../../../../domain/types/Project";
 import useProjectData from "../../../../../application/hooks/useProjectData";
-import CustomizableSkeleton from "../../../../vw-v2-components/Skeletons";
+import CustomizableSkeleton from "../../../../components/Skeletons";
 import { formatDate } from "../../../../tools/isoDateToString";
-import {  useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import { User } from "../../../../../domain/types/User";
 import { getEntityById } from "../../../../../application/repository/entity.repository";
 import useProjectRisks from "../../../../../application/hooks/useProjectRisks";
